pyproject.toml
[project]
name = "python-automation-hw-19"
version = "0.1.0"
description = ""
authors = [
    {name = "Filipp Goncharenko",email = "phil.can@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pytest (==8.4.1)",
    "selene (==2.0.0rc9)",
    "appium-python-client (==5.1.3)",
    "allure-pytest (==2.15.0)",
    "requests (==2.32.4)",
    "python-dotenv (==1.1.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
